MapItel/requirements.txt
newsapi-python
openai~=0.27.2
selenium~=4.8.2
requests~=2.28.2
beautifulsoup4~=4.11.2
Flask~=2.2.3
tqdm~=4.65.0
pandas~=1.5.3